fix: allow fractional SVG view dimensions

+++ server/validation.mjs
@@ -55,7 +55,7 @@ export const schemas = {
       // de corpo do Express (20 MB) para o JSON inteiro caber.
       .max(15_000_000, 'SVG muito grande (máx. ~15 MB)')
       .refine((m) => m.trimStart().startsWith('<svg'), 'markup deve ser um SVG'),
-    viewW: z.coerce.number().int().positive().max(100_000).optional(),
-    viewH: z.coerce.number().int().positive().max(100_000).optional(),
+    viewW: z.coerce.number().positive().max(100_000).optional(),
+    viewH: z.coerce.number().positive().max(100_000).optional(),
   }),
 };
